minor wsi tests cleanup
